backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Authlib==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
